Combine lost features on subscription end in one component
